fixed masked boundind-box index bug tested albedo derivatives

+++ src/SparseMatrix.cu
@@ -4,14 +4,14 @@ rtBuffer<float3, 1> values;
 rtBuffer<unsigned int, 1> rowOffsets;
 rtBuffer<unsigned int, 1> colIndices;
 
-RT_CALLABLE_PROGRAM void Add(size_t row, size_t col, float3 value)
+RT_CALLABLE_PROGRAM void Add(uint row, uint col, float3 value)
 {
-  size_t begin = rowOffsets[row];
-  size_t end = rowOffsets[row + 1];
+  unsigned int begin = rowOffsets[row];
+  unsigned int end = rowOffsets[row + 1];
 
   if (begin < end)
   {
-    size_t index = begin;
+    unsigned int index = begin;
 
     while (begin < end)
     {
